backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
websockets
python-dotenv
httpx
aiohttp
